added nice interactive output for refresh script (plus purge data on PURGE_DATA=)

+++ refresh.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
-rm -rf ~/{.local/{share,state},.cache}/nvim
+function refresh_nvim() {
+    [[ -v "PURGE_DATA" ]] && rm -rf ~/{.local/{share,state},.cache}/nvim
 
-nvim --headless "+StarterPack" \
-    "+lua require('nvim-treesitter.configs').setup { ensure_installed = {       \
+    nvim --headless "+StarterPack" \
+        "+lua require('nvim-treesitter.configs').setup { ensure_installed = {       \
     'bash',             \
     'cpp',              \
     'css',              \
@@ -24,3 +25,18 @@ nvim --headless "+StarterPack" \
     'typescript',       \
     'yaml',             \
     }, sync_install = true }" "+qa"
+}
+
+if [[ -v "INLINE_OUTPUT" ]]; then
+    tput rmam
+    refresh_nvim 2>&1 | while read -r line; do
+        if [[ -n "$line" ]]; then
+            echo -ne "\r\033[2K"
+            echo -n $line
+        fi
+    done
+    echo -ne "\r\033[2K"
+    tput smam
+else
+    refresh_nvim
+fi
